- Added a alert when the user try to delete a expense; - Layout adjustments;

+++ ExpensesControlApp/ExpenseDetail.xaml.cs
@@ -13,11 +13,26 @@ namespace ExpensesControlApp
             InitializeComponent();
         }
 
+        protected override async void OnAppearing()
+        {
+            base.OnAppearing();
+
+            var expense = (Expense)BindingContext;
+            title.Text = expense.expenseTitle;
+            description.Text = expense.expenseDescription;
+            amount.Text = expense.expenseAmount;
+        }
+
         async void didTapDeleteExpense(object sender, EventArgs e)
         {
-            var expense = (Expense)BindingContext;
-            await App.Database.DeleteExpenseAsync(expense);
-            await Navigation.PopAsync();
+
+            bool confirmCancel = await DisplayAlert("Atenção", "Realmente deseja deletar essa despesa?", "Sim", "Não");
+            if (confirmCancel)
+            {
+                var expense = (Expense)BindingContext;
+                await App.Database.DeleteExpenseAsync(expense);
+                await Navigation.PopAsync();
+            }
         }
 
         async void didTapEditExpense(object sender, EventArgs e)
@@ -27,8 +42,6 @@ namespace ExpensesControlApp
             {
                 BindingContext = expense
             });
-
-            await Navigation.PopToRootAsync();
         }
     }
 }
